Customize common urls (#1728)
## Context

The grist gouv team often has the need to customize the custom URLs.
Thus we open PRs each time.

This can be a bit tedious, long before we have it in production and we
already have plenty URL customisations:
 - GRIST_HELP_CENTER
 - GRIST_ONBOARDING_VIDEO_ID
 - GRIST_TERMS_OF_SERVICE_URL
 - FREE_COACHING_CALL_URL (hmm, without prefix, it is not consistent)
 - GRIST_CONTACT_SUPPORT_URL
 - GRIST_WEBINARS_URL
 - GRIST_OPEN_GRAPH_PREVIEW_IMAGE

## Proposed solution

I propose to rather introduce a new env variable:
`GRIST_CUSTOM_COMMON_URLS`, which is expected to be a valid JSON object.

Each keys are documented in the interface descriptor that is moved in a
separate module.

## Has this been tested?

- [x] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->