(custom widget) disable space key shortcut when necessary
before this commit, we could not press the space key to trigger the
"open configuration" button of a custom widget (which is in the right
panel). this was because the viewAsCard command, mapped on the space
key, was always enabled.
Now the command is enabled only when we focus the widget, allowing to
correctly use the space key in the right panel for that specific case.