(right panel) sort item dots menu: prevent infinite loop bug
it seems there is a bug in weasel when a menu is composed of
non-selectable items. If we open a sort item dots menu and try to use
the arrow down/up keys, it goes into an infinite loop.

As a workaround for now, prevent keyboard selection. Not ideal but at
least it doesn't crash.