Adds a list of users currently accessing the doc (#1739)
## Context

It can be useful when editing a doc to see who else is currently accessing that document. 

For example, to check everyone on in a call has the document open, or to avoid concurrent editing.

The tooling for basic user presence also enables potential future expansions, such as sharing editors' cursor positions.

## Proposed solution

This adds a list of users to the top bar, that's updated whenever someone opens the document in a new tab or window.

The is tracked on a per-session basis, with multiple tabs for a single user resulting in multiple entries.