(core) Fix regression with tooltip close
Summary:
Pointer events were not suitable for some logic in
hover tooltips that defer closing until a user finishes
selecting text in them. This was causing a bug where
a tooltip would not close on pointerleave because it
was stuck in a mouse grabbed state, even though a
click never occurred, since pointerdown can trigger
without a click.

Test Plan: Existing tests

Reviewers: paulfitz

Reviewed By: paulfitz

Subscribers: paulfitz

Differential Revision: https://phab.getgrist.com/D4695