(chartview) fix chart field remove icon not working when not in English
funny bug: the 'Remove' icon codename was wrongfully passed to the
translation system. When using Grist in French for example, this
resulted in the system trying to find a "Retirer" icon (French
translation for "Remove"). That obviously didn't work and made a gray
square appear in place of a bin icon.