CensorshipInfo: remove apply, move logic into filter
Apply and filter are basically the same thing except filter can work
`_grist_cells` but apply can't. The code as written also has an
assumption that any table that is not in `this.censored` is an ACL
table, an assumption we're about to break.

Instead, be more explicit about the three tables that may need ACLs
wiped out and also be explicit that only tables in `this.censored` can
be censored.