Adds browser tests for the "Active user list" user presence feature (#1788)
## Context

The initial commit for the active user list had backend tests only.

## This commit

This adds browser tests for the frontend rendering, that complement the
networking and privacy checks run on the backend.

It also adds a GRIST_USER_PRESENCE_MAX_USERS parameter to limit the number of active users displayed simultaneously.

## Related issues

#1757 

## Has this been tested?

- [X] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->