update endpoints for deleting sites and workspaces (#1738)
The easiest-to-find endpoints for deleting sites and workspaces were
old, from an early Grist prototype. The correct method for deleting
sites in particular was hard to find, hidden in a `Doom` class. This
updates the endpoints to use the preferred implementation.

Note that these endpoints are very unforgiving. Call them to delete a
resource, and it ends up deleted, with no soft-delete step. There is a
`/remove` endpoint for workspaces (and documents) that does
soft-deletes.

For the org delete endpoint, there is now an extra `/{name}` path
segment added at the end of the URL, which must match either `org.name`
or `org.domain` or `force-delete`. This is a breaking change. To
continue using the endpoint without this new path segment, set
`GRIST_ALLOW_DEPRECATED_BARE_ORG_DELETE=1` in the environment.