fix: Missing translations in Admin panel (#1524)
## Context

Some strings are not translated in AdminPanel.

* Some strings are not prefixed with the good scope
* Some strings are not passed to i18n t() function in the code

## Proposed solution

This PR only does partially it only fixes client side strings.

## Out of scope

Some more job has to be done as some chains are generated server side,
at least in `app/server/lib/BootProbes.ts`
And it seems to use some keys of objects as string to be translated,
which will need some more job.

## Has this been tested?

- [x] 👍 yes, new strings are in `static/locales/en.client.json