Dockerfile: install devDependencies on top of production dependencies (#1812)
## Context

We want the final Docker image to only contain the NodeJS dependencies
required to run the server, i.e. exclude those marked as
`devDependencies` in `package.json`. However we need the
`devDependencies` to build the project.

The current way we achieve this is that in the `builder` stage of the
`Dockerfile`, we run `yarn install` once to install all dependencies to
`node_modules`, then run it a second time to install only the production
dependencies (using the `--prod` flag to `yarn install`) into a separate
`node_modules_prod` directory. Then in the final stage we copy the
`node_modules_prod` directory from the `builder` stage to be the final
`node_modules`.

An issue with installing packages to a non-standard location such as
`node_modules_prod` is that packages that have `install` scripts are not
able to resolve dependencies they need as part of their installation
step (e.g. to build native modules), because those scripts will still
look for modules in the standard `node_modules` directory.

A workaround was developed that involved globally installing these
dependencies, but this is fragile as we have no control over which
dependencies any package may require as part of its installation.

## Proposed solution

Instead of installing to `node_modules_prod`, we install the production
dependencies to the standard `node_modules` location in an initial
`prod-builder` stage.

Then in the `builder` stage which is based on the `prod-builder` stage
we run `yarn install` again without the `--prod` argument, which will
cause it to install the additional dependencies (without downloading
anything, as the Yarn cache is preserved from the previous `yarn`
invocation).

In the final image, we copy the `node_modules` directory from
`prod-builder`, ensuring we only pick up the production dependencies.

## Impact

This change could make the Docker image build slightly faster, but the
main benefit is to no longer have a list of arcane build-time
dependencies such as `node-pre-gyp` (potentially with the wrong
versions, too) in the `Dockerfile`.

The final image should remain essentially identical.

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [x] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->