add a placeholder for proposed changes UI
This adds a placeholder for a "crowdsourcing" UI. It is very experimental,
and not worth you looking at yet.

Main changes:

  * Add a "Proposed Changes" page on forks that shows their
    difference from the trunk. Uses the mechanism for showing
    changes that is used by Document Activity tab, which is
    VERY CLUNKY AND UGLY.
  * Add a little button to tabular diffs to show unchanged context
    with them. This is just for experimenting so far and currently
    adds WAY TOO MUCH.

Mostly some refactoring so far. The actual logic used for computing context
is a bit wrong, fragile, and incorrect - it is just a placeholder.

Part of this project: https://github.com/gristlabs/grist-core/issues/1829