Forms: Rephrase the tooltip to reject other instance hosters responsi… (#1810)
## Context

The forms display a tooltip with this content:

<img width="671" height="236" alt="image"
src="https://github.com/user-attachments/assets/6b5e8ec6-1340-4ae7-bbf9-e6d7c955ff75"
/>

Textual content: 
> This form is created by a Grist user, and is not endorsed by Grist
Labs. Do not submit passwords through this form, and be careful with
links in it. Report malicious forms to
[support@getgrist.com](mailto:support@getgrist.com).

But it does not tell anything about other service providers or hosters
responsibility. It would be clearer to reject theirs too.

## Proposed solution

Rephrase this tooltip with this content :
> This form is created by a Grist user, and is not endorsed by the
editor (Grist Labs) nor by any party providing this service. Do not
submit passwords through this form, and be careful with links in it.
Report malicious forms to
[support@getgrist.com](mailto:support@getgrist.com).

## Has this been tested?

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [x] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->

---------

Co-authored-by: nbush <njdbush@gmail.com>