add instructions on how to build with grist-ee (#1826)
There is demand for Grist Enterprise among people building from source.
This adds instructions for how to do so.

The `WEBPACK_EXTRA_MODULE_PATHS` bit wasn't necessary in the past, I
think there was a little extra logic in the webpack config file, but
maybe I messed it up when working on grist-desktop?