(core) Background assistant while popup is closed
Summary:
The assistant popup is no longer disposed when closed,
which means that responses continue to be processed in
the background, and state such as scroll position are
preserved.

Additionally, structured output is now disabled by default.
A new environment variable can be set to enable it, which
is only intended to be set while running AI completion tests.
This should improve compatibility with LLM endpoints that
don't support structured output, and should generally reduce
latency outside of tests.

Finally, we now hide the comments announcement if we're
in the process of starting a new document with the assistant.

Test Plan:
 - New browser test
 - Existing tests

Reviewers: jarek

Reviewed By: jarek

Differential Revision: https://phab.getgrist.com/D4722