(core) Fix issue with View-As simulation getting incorrect 'access' property
Summary:
In some cases, such as for CSV exports, and depending on the rules, "View As"
was only partially respected, and rules depending on user.Access were simulated
using the actual owner's access rather than that of the "View As" user.

Make slightly more concise and hopefully clearer

Test Plan:
Added some asserts to the existing ExportsAccessRules test case to
ensure these are respected. The test case fails without this fix.

Reviewers: georgegevoian

Reviewed By: georgegevoian

Subscribers: georgegevoian

Differential Revision: https://phab.getgrist.com/D4718