Prevent active user list closing on change (#1789)
## Context

When more than 4 users have a document open, the active user list shows a "remaining users" button. This button opens a popup menu with a list of all users in.

When a session is created / destroyed (i.e. another tab is opened with the document in), the user list will close automatically, forcing the user to re-open it.

## This commit

This prevents the list being closed by restructuring how the user list HTML is computed, allowing the contents to be updated without replacing the whole list.

## Related issues

#1739