add a timeout to broadcast messages
We saw an instance of a document hanging on shutdown because
broadcasts never terminated. This adds a timeout if some client
is hanging. This does add some risk that a slow broadcast that
might have succeeded in the past will now fail. The timeout is
kept quite high (60 seconds), and a problem connection is closed
to avoid state getting weird.