add a placeholder for "Proposed Changes" UI (#1837)
This adds a page to forked documents to show their differences from the
trunk using the same format used to show document activity (the
`ActionLog` component).

How to test:

  * Log in and make a document with a table with some content.
* From the share menu, select "Work on a Copy" and then make some change
to it.
  * In the left sidebar, there should be a new "Proposed Changes" page.

<img width="1006" height="552" alt="Screenshot from 2025-09-17 17-15-37"
src="https://github.com/user-attachments/assets/819e9fb4-ed11-498d-b6a8-de87f1001f23"
/>

The ActionLog format is really bad! Incomplete, out of order, buggy.
This work does not fix that. It moves some of the code around without
fixing it or improving it. Yet.

This is a placeholder for work on crowdsourcing, where this page will
also appear in trunk documents, showing any contributions from forks
that the fork owner has chosen to "propose". That step of the work is
not included.

There is one change made to the representation of diffs. A little ">" is
shown beside tables which, if pressed, adds more context information to
them. Currently this adds all current cells in the rows shown in the
diff. This is experimental work just to test how practical it is to
access this data.

Part of this project:
https://github.com/gristlabs/grist-core/issues/1829