(core) Upgrade tmp-promise and unify a few scattered versions of the tmp dependency
Summary:
There was a discrepancy between tmp's behavior, and that of the older
tmp-promise (which wrapped an older tmp), which manifested as inability to
import files when the temp directory involved a symlink.

Updating tmp-promise makes it use the same more recent version of tmp, which
fixes the issue.

Test Plan: Tested manually an import on mac.

Reviewers: jarek

Reviewed By: jarek

Differential Revision: https://phab.getgrist.com/D4728