Accessibility options modal (#1734)
## Context

Now that there is a high contrast theme, and most importantly, that
keyboard shortcuts exist to escape the widget and navigate through the
whole document page ; it feels like a good idea to expose that more in
the UI.

## Proposed solution

The idea is to have an "Accessibility" button directly available in the
tools section of the left panel.

- this button has a rather common icon for accessibility related things
on the web ; currently borrowed from bootstrap icons as an example (I
didn't get if an icon lib was already in use or if icons were homemade),
- we directly show in the UI the keyboard shortcut triggering this
button (F4 for now, _but I need more testing on this_). I feel this is
important: since by default, we can't access this button by normal usage
of the tab key when on a document page, visually showing the
corresponding shortcut allows to trigger the button even if we don't
know about the region switching shortcuts
- when clicking the button, we see a modal, letting us quickly enabling
the high contrast theme, and explaining us the region switching keyboard
shortcuts exist

## Has this been tested?

- [x] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->

## Screenshots / Screencasts


https://github.com/user-attachments/assets/81b951c4-fdd2-4482-b67c-f9b68c8908c8