Log the result of the `/status` endpoint (#1858)
## Context

The /status endpoint is usually used as probes in Kubernetes. In case
this endpoint returns a result suggesting it is unhealthy, it is more
convenient to be able to also read the result using the app logs.

## Proposed solution

A new log message is logged for every failure, and every 100 successes, and includes the details of the status checks.

This also removes an existing log message about status requests, which produced a lot of logging but little benefit.

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [x] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->