(searchbar) fix searchbar closing when switching pages on search
when checking "search on all pages", the searchbar closed when
switching page because it listened for the RFS changes and called
`toggleMenu(false)`.

Change the logic so that the searchbar stays open on page changes
if we detect that it was opened before, + add a custom outside clicks
listener to properly handle closing the searchbar when we dont want it
anymore.