(searchbar) expose a new Mod+Enter shortcut to quickly get out
This is already doable with escape… But. Before, we could just press Tab
to exit the searchbar easily. Now, pressing Tab correctly tabs through
items next to the searchbar (like the "search all" checkbox). So this
quick way of getting out the searchbar is gone.

I felt like exposing a "submit search" action through Mod+Enter was a
good idea, it seems it better fits the intent than "escaping" search
bar. But I guess this is a bit nitpicking, we could also do without
this?