Handle better long list in forms #1144 (#1781)
## Context

Previously the list for choices and references were truncated to the
first 1000 options and the search a bit tedious (the results were not
sorted by relevance so we could not see the best matches by a single
look).

## Proposed solution

This commit brings the following:
1. The result are sorted by relevance just like in the grid view
2. The fake "Select..." option is removed, as it would have otherwise
been sorted. It used to empty the field. Instead add a cross button
(accessible via the keyboard navigation) that appears when a value is
selected.
3. Reduce the number of displayed elements
4. Remove the limit of candidate options in the list (i.e. options that
are not displayed but searchable and selectable)
5. Show an indicator message at the end of the list ("Showing M of N items. Search for more")

Further proposed improvements:
1. Being able to match items that match searched string, even when it
does not start with the user input (like `bar` should match `foobar`).
But this PR is meant to *first and foremost* make these lists behave
just like their equivalent in the Grid (as much as possible). I consider
that part out of scope at this step.
2. (Technical): Factorize code between `RefRenderer` and
`ChoiceRenderer`? (Maybe that's harder that it sounds?)

## Related issues

Partially fixes #1144, other improvements will be made in a futur PR.

## Has this been tested?

- [ ] 👍 yes, I added tests to the test suite
- [x] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->

## Screenshots / Screencasts

### Before this PR

(Please note that in this video `Foo17` appears at the top in the long
ref list because I made test, sorry for the confusion)



https://github.com/user-attachments/assets/bc395d9b-492c-4ddb-b9da-6c876325cf84



### With this PR


https://github.com/user-attachments/assets/5fc872f1-c340-47e1-b5eb-b6157036c250