Make searchbar more keyboard-friendly (#1752)
- (kb) make sure ViewLayout escape command don't collapse with others

This is a workaround exposing an issue that can happen with keyboard
handling:

When the RegionFocusSwitcher (RFS) focuses the ViewLayout, the
ViewLayout registers its commands again, making those commands the
"current" commands attached to the keyboard shortcuts.

So, if before focusing the ViewLayout, we had registered commands for
stuff that we would like to actually keep as "current" even after
focusing back the ViewLayout, we are not happy, because they get
overriden by the ViewLayout commands.

Here we decide to manually handle the escape key of the ViewLayout so
that it's only registered when we detect a maximized leaf, to fix a
specific issue related to this.

The issue was:

- I select cells in a grids. cancel command is now clearSelection from
the GridView
- I then open the searchbar. RFS focuses top panel,
- I then press Esc. The searchbar closes, RFS focuses the ViewLayout,
making ViewLayout commands be the current ones, including the "cancel"
one
- I then press Esc again, expecting to clear my selection. I can't
because the current cancel command is the ViewLayout's, basically doing
nothing when not maximized.

Now:

- I select cells in a grids. cancel command is now clearSelection from
the GridView)
- I then open the searchbar. RFS focuses top panel,
- I then press Esc. The searchbar closes, RFS focuses the ViewLayout,
making ViewLayout commands be the current ones, except for the "cancel"
one, because it activates only when the view is maximized,
- I then press Esc again, expecting to clear my selection, and my
selection is correctly cleared.