add download menu option to document list (#1843)
This is easier to get to if the document is in a bad state. The endpoint
will still attempt to open the document, but at least there's no need to
render the web client. And the endpoint allows downloads to owners even
if granular access control gets messed up.

## Context

There are several ways to get a document into a bad state. Once in a bad
state, the web client may not function correctly, which currently could
impede access to the UI for downloading a document. There's a somewhat
robust endpoint for downloading, but it would be nice to have reliably
available UI.

## Proposed solution

Add a download option for documents in listings, not just in the share
menu when they are open.

## Related issues

https://github.com/gristlabs/grist-core/issues/1784

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [x] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->

## Screenshots / Screencasts

<img width="865" height="553" alt="Screenshot from 2025-09-30 11-11-34"
src="https://github.com/user-attachments/assets/d4ce72ab-a0cb-43a4-99c4-ab181ef5ad45"
/>