Refactor mobile button detection logic (#1862)
Currently Grist detects mobile browser by looking at the `UserAgent`
browser settings, but some browsers
hide that for security reasons.

## Proposed solution

Instead of detecting the mobile platform, Grist will now check what is
the primary input device or if
the screen is small enough to assume it is a mobile device.

## Has this been tested?

Only manual tests on mobile browsers