remove height/align config when transforming a column to choice type
In the right panel, when setting the column type to "choice" or "choice
list", in the popup appearing to configure the new type, we can edit the
choices (good).

But above that, we see that we can also edit the text alignment of the
column, and since recently, also the whole row height.
This is more questionable, I'd argue it's stuff getting in the way as I
guess we'd mostly want to edit the choices at this moment, and not
bother with text alignment or row height. Especially row height, that
triggers a tab change to the "table" tab of the creator panel.