(chart) add space between each slice of a pie chart
For now, pie chart slices are only differentiated with slice colors.
That can make the data a bit troublesome to visualize for people having
vision impairments, where they could not easily understand where one
slice is stopped or starts, especially when two slices or a similar
color are next to each other.

This marks every slice with borders to prevent those questions.