(chart) make pie chart slices text more pronounced
For now, pie chart text on slices is sometimes hard to read. The
contrast ratio with the background color can be a bit low. To consider a
thing readable, we should target a 4.5:1 contrast ratio with text and
bgs in general. Here, plotly makes sure to use light or dark text
automatically depending on slice color, but it's not always a high
enough contrast ratio to match our requirements.

Fixing that correctly might be a bit too much work for now compared to
other things to work on. As a "quick win", this commit makes the pie
slices text a bit bigger, and most importantly, make it bold. This
already makes the text noticably easier to read compared to before.