Split member retrieval query from main document query
The subqueries generated by `HomeDBManager._markIsPermitted`
are re-executed for each result row if the `docs` table is joined with
another (the PostgreSQL optimizer is not smart enough to realize
that since the document ID is the same on every row of the result, it
could do each subquery just once).

Since each of these subqueries scans the users attached to the
document via `acl_rules` itself, the cost of a query that includes
these subqueries and also retrieves the document members
via a join to `acl_rules` grows quadratically with the number
users attached to document.

Instead what we do now is to split the retrieval of the document
member list from the initial document query, as was already
applied for the workspace and organization members.