fix(MinioExternalStorage): ECONNRESET and InternalError are fatalErrors (#1874)
## Context

https://github.com/gristlabs/grist-core/commit/e6692c279332d3e6a168df5d1f304c38da6d0dde
mistakenly added 2 error codes to the `isfatalError` method which acts
as a sort of whitelist for MinIO errors that should be safe to ignore.
ECONNRESET and InternalErrors **are** fatal errors.

In production, this meant that when our MinIO server was experiencing
restarts or overall hiccups, the MinioExternalStorage could create a new
document when one already existed, and then push it to MinIO,
overwritting the existing document on the bucket 🥵

## Proposed solution

Remove ECONNRESET and InternalErrors from the list of error codes
MinioExternalStorage should ignore.

## Related issues

Wondering whether the naming of the method lead to confusion in the
first place, should we maybe rename it to `isNotExistingError` or
`isExpectedError` ? I feel like the logic of the code in `head` would
read better (although the comment I added should hopefully fix it).

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [x] 🙋 no, because I need help

I'm not sure whether I need to add a test here, but apparently the
faulty code was added in order to fix test-related weirdness, so hoping
the tests will pass now.