(ui) restyle button groups to better distinguish active element
button groups/button toggles like:

- TextBox/Spinner choice in column config when the column is a number
type,
- or when selecting text alignment in column config,
- or the text wrap toggle button in column config,

had styles that may be confusing. Especially when we had groups of 2
buttons, or for the toggle button, we could have trouble understanding
what was the active button/state and what was the inactive one. Because
we "only" changed the main color and it may not be obvious what color
meant "active" and what color meant "inactive".

New styling add borders around the active button/state so that we 100%
understand the difference.

Plus, instead of using a black color that was mostly used only for these
buttons, we rely on the primary color palette to have a more consistent
UI with other components.