Fixes /copy endpoints not working with short url ids (#1881)
## Context 
`/docs/:docid/copy` fails when using the url id (the short identifier in the document's URL), when multiple doc workers are in use.

The following explanation is the most likely reason, but hasn't been confirmed through empirical testing.

When the copy request is made, it's assigned to a new doc worker to handle the copy and import of the new document. That doc worker first attempts to download the original document. This download is either from a doc worker that currently has that document assigned, or a newly assigned doc worker.

The copying doc worker attempts to lookup the doc worker to download from using its short id. This instead assigns a new doc worker, as doc worker lookups should only ever use the document's full id.

This results in the download throwing an error, because the document is already assigned to another worker.

## This commit

This changes `fetchDoc` to resolve the full document id first, so that the doc worker lookup / assignment functions correctly.