(core) remove the support user from the set of admins when possible
Summary:
When setting up Grist, the admin page is useful for diagnosing problems. But to view it you need admin rights, and who is admin is a bit confusing in Grist currently.

In grist-ee, a `GRIST_INSTALL_ADMIN_ORG` variable is supported, which - if set - limits admins to exactly the owners of the specified org. We leave that unchanged.

Otherwise, the admins until now are the user specified by `GRIST_DEFAULT_EMAIL` if set, and the support user (which defaults to `support@getgrist.com`, overridable with `GRIST_SUPPORT_EMAIL`).

The support user was added in order to have a predictable admin user, for use with boot keys. But this isn't great, since an administrator may not be aware of the support user. So this change modifies boot keys to use the admin user specified by `GRIST_DEFAULT_EMAIL` if set. The support user is now only admin if `GRIST_DEFAULT_EMAIL` is not set.

For better or worse, in the standard startup code for Grist (`stubs/app/server/server.ts`), a default email of `you@example.com` is set.

Boot keys didn't previously work when `GRIST_INSTALL_ADMIN_ORG` is set, since they would authenticate the user as the support user, who was not necessarily an admin. Now they authenticate as one of the admin users picked via an arbitrary rule (lowest id).

Test Plan: Tests are updated. GRIST_INSTALL_ADMIN_ORG tested manually.

Reviewers: jarek

Reviewed By: jarek

Subscribers: jarek

Differential Revision: https://phab.getgrist.com/D4749