Fix Firefox ESR blocking bug in comments UI (#1884)
## Context

The before-latest firefox ESR version (128, released in 2024) does not
support `contenteditable="plaintext-only"`.

This makes adding comments and replying to comments impossible when
using Firefox ESR 128, as this contenteditable value is used in the
comments UI.

This bug was reported to us a few times because some French
administrations use Firefox ESR. There might be quite a while before
they update their ESR version, so I figured a tiny fix here could
greatly help.

## Proposed solution

This adds a tiny support check for `contenteditable="plaintext-only"` in
the comments UI. If we detect a lack of support, we manually clean the
clipboard content before pasting it, in case we copy formatted text in
the comment box. This basically mimics `plaintext-only` behavior.

## Has this been tested?

This has been manually tested on Firefox 128. This can be tested by
downloading Firefox 128 ESR, or any Firefox 128 version. For example on
Linux you can download an old Firefox Developer Edition version that
doesn't get in the way of your normal Firefox install.

- [x] 🙅 no, because this is not relevant here

---------

Co-authored-by: Florent <florent.git@zeteo.me>